Move admin tests to a different folder
